+++ src/Presentation/SmartStore.Web.Framework/UI/NotifyType.cs
@@ -1,9 +1,0 @@
-﻿namespace SmartStore.Web.Framework.UI
-{
-    public enum NotifyType
-    {
-        Success,
-        Error,
-        Info
-    }
-}